# web-app-backend/src/main/java/com/swozo/api/web/activity/dto/ServiceConnectionDetailsDto.java
package com.swozo.api.web.activity.dto;

import com.swozo.model.utils.InstructionDto;
import com.swozo.utils.SupportedLanguage;
import io.swagger.v3.oas.annotations.media.Schema;

import java.util.Map;
import java.util.Optional;

public record ServiceConnectionDetailsDto(
        @Schema(required = true) Map<SupportedLanguage, InstructionDto> connectionInstructions,
        @Schema(required = true) String url,
        Optional<String> connectionInfo
) {
}